fix: avoid narrow-dtype overflow in reduceat group offsets
_calc_counts_invidx cast its inv_idx/counts arrays to groups.dtype, the
dtype pydata/sparse picks to fit the array's *shape*. But those arrays
are used as offsets into the nnz-length data array via
np.add.reduceat, which can be far larger than any shape dimension.
Once nnz exceeded the dtype's max (e.g. 32767 for int16, a legitimate
choice for small-shaped arrays), the offsets silently wrapped negative,
raising "IndexError: index ... out-of-bounds in add.reduceat" from
COO.sum/.mean/.max and friends.

Hardcode int64 for these two transient arrays instead of reusing
groups.dtype -- they are only used as index arguments, never stored on
the array, so the memory-saving rationale for a narrow dtype doesn't
apply to them.

+++ sparse/numba_backend/_coo/core.py
@@ -1600,13 +1600,18 @@ def as_coo(x, shape=None, fill_value=None, idx_dtype=None):
 
 @numba.jit(nopython=True, nogil=True)  # pragma: no cover
 def _calc_counts_invidx(groups):
+    # NB: inv_idx/counts index into the (potentially much larger) nnz-length data
+    # array, so they must not be narrowed to groups.dtype, which is only sized to fit
+    # the group values (e.g. a shape dimension). Doing so silently overflows for
+    # arrays with more stored elements than the dtype max, e.g. int16 wraps negative
+    # past 32767 nnz, causing an IndexError in the np.add.reduceat call downstream.
     inv_idx = []
     counts = []
 
     if len(groups) == 0:
         return (
-            np.array(inv_idx, dtype=groups.dtype),
-            np.array(counts, dtype=groups.dtype),
+            np.array(inv_idx, dtype=np.int64),
+            np.array(counts, dtype=np.int64),
         )
 
     inv_idx.append(0)
@@ -1620,7 +1625,7 @@ def _calc_counts_invidx(groups):
 
     counts.append(len(groups) - inv_idx[-1])
 
-    return (np.array(inv_idx, dtype=groups.dtype), np.array(counts, dtype=groups.dtype))
+    return (np.array(inv_idx, dtype=np.int64), np.array(counts, dtype=np.int64))
 
 
 def _grouped_reduce(x, groups, method, **kwargs):
